KAFKAC-10: Implement partial initialization of the stream state
Motivation
----------
It is useful to assign connector to watch only for subset of
partitions (vbuckets). So that connectors could be organized in a group
and workload will be balanced.

Modification
------------
Adopt new BucketStreamAggregatorState API

Result
------
It is possible to select which partitions will be watched and replicated
to kafka

Reviewed-on: http://review.couchbase.org/56075
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ src/main/java/com/couchbase/kafka/state/StateSerializer.java
@@ -9,9 +9,9 @@ import com.couchbase.client.core.dcp.BucketStreamState;
 public interface StateSerializer {
     void dump(BucketStreamAggregatorState aggregatorState);
 
-    void dump(BucketStreamAggregatorState aggregatorState, int partition);
+    void dump(BucketStreamAggregatorState aggregatorState, short partition);
 
     BucketStreamAggregatorState load(BucketStreamAggregatorState aggregatorState);
 
-    BucketStreamState load(BucketStreamAggregatorState aggregatorState, int partition);
+    BucketStreamState load(BucketStreamAggregatorState aggregatorState, short partition);
 }
